pyproject.toml
[tool.poetry]
name = "Polylist"
version = "0.1.0"
description = ""
authors = ["Mohammed Salman <mohamed.as90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-fasthtml = "^0.4.5"
pytest = "^8.3.2"
beautifulsoup4 = "^4.12.3"
langchain = "^0.2.15"
langchain-openai = "^0.1.23"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.15"
pytest-asyncio = "^0.24.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
